# services/api/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Shipped API
  version: "1.0.0-phase2"
  description: >
    The Shipped control plane. Most /v1 routes require a Bearer EdDSA JWT
    (issued by Better Auth) which the Go API verifies itself (the authz boundary).
    Every request runs under a Postgres RLS tenant context derived from the
    verified org_id/user_id claims.


    Phase 2 adds access control & domains: the /authz viewer exchange (mint +
    password) which issues a host-scoped EdDSA edge token the serving Worker
    verifies (public JWKS at /.well-known/edge-jwks), per-site access-mode + policy
    management (admin/owner only), allowlist CRUD, the org allow_external_sharing
    toggle (with reconcile), member listing, and Cloudflare-for-SaaS custom domains.
  license:
    name: FSL-1.1-Apache-2.0
servers:
  - url: https://api.shipped.app
    description: Production
  - url: http://localhost:8080
    description: Local / self-host
security:
  - bearerAuth: []
paths:
...
  /v1/members:
    get:
      summary: List the caller org's members
      description: >
        Lists members from the Better Auth `member` table, RLS/org-scoped. The Go
        API only READS roles (Better Auth owns invite/accept/role mutations). If the
        Better Auth table is unavailable (self-host pre-Better-Auth), returns an
        empty list.
      operationId: listMembers
      responses:
        "200":
          description: The org's members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items: { $ref: "#/components/schemas/Member" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/orgs/allow-external-sharing:
    put:
      summary: Toggle the org allow_external_sharing policy (admin/owner only)
      description: >
        Owner/admin only (role re-checked against the member table). When DISABLING,
        reconciles: public sites are downgraded to org_only and their edge routes
        rewritten; external-email allowlist grants are revoked. Enabling only widens
        what is permitted (no reconcile).
      operationId: setAllowExternalSharing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: The new policy + the number of downgraded sites
          content:
            application/json:
              schema:
                type: object
                properties:
                  allow_external_sharing: { type: boolean }
                  downgraded_sites: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /v1/sites:
    post:
      summary: Create a site
      description: >
        Reserved slugs are rejected (400). A slug whose content host
        (`<slug>.shippedusercontent.com`) is already taken by another org/site
        returns 409 Conflict — the host registry is global even though site slugs
        are unique only per org. On the cloud build, exceeding the per-user site
        cap returns 402 with the quota body so the dashboard can open the upgrade
        modal.
      operationId: createSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:
                  type: string
                  description: Subdomain label under shippedusercontent.com (unique per org).
                  example: my-docs
      responses:
        "201":
          description: Site created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/QuotaExceeded" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      summary: List the caller org's sites
      operationId: listSites
      responses:
        "200":
          description: The org's sites
          content:
            application/json:
              schema:
                type: object
                properties:
                  sites:
                    type: array
                    items: { $ref: "#/components/schemas/Site" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/sites/{id}/access:
    put:
      summary: Set a site's access mode + policy (admin/owner only)
      description: >
        Owner/admin only (role re-checked against the member table). Sets the
        access_mode (public/password/allowlist/org_only) and the matching policy:
        a password mode's password is hashed server-side (bcrypt; plaintext is never
        stored), optional expires_at (RFC3339) sets a link expiry, unlisted marks a
        public-tier site as unlisted. Rewrites the edge RouteValue (mode + expires_at
        for the public tier). A public site under allow_external_sharing=false → 403.
      operationId: setSiteAccess
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode: { type: string, enum: [public, password, allowlist, org_only] }
                password: { type: string, description: Plaintext; hashed server-side. Required for mode=password. }
                expires_at: { type: string, format: date-time, description: Optional link expiry (RFC3339). }
                unlisted: { type: boolean, description: Public-tier unlisted flag. }
      responses:
        "200":
          description: Access updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  site_id: { type: string, format: uuid }
                  mode: { type: string }
                  unlisted: { type: boolean }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sites/{id}/allowlist:
    get:
      summary: List a site's allowlist
      operationId: listAllowlist
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      responses:
        "200":
          description: The site's allowlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowlist:
                    type: array
                    items: { $ref: "#/components/schemas/AllowlistEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: Add an email to a site's allowlist (admin/owner only)
      description: >
        Owner/admin only. is_external is set automatically when the email's domain is
        not one of the org's VERIFIED domains; the DB trigger then rejects an external
        grant under allow_external_sharing=false (403).
      operationId: addAllowlistEntry
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "201":
          description: Entry added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AllowlistEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Remove an email from a site's allowlist (admin/owner only)
      operationId: removeAllowlistEntry
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Entry removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived EdDSA JWT (Better Auth) or a deploy token.
  parameters:
    SiteID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: app.sites.id
  schemas:
...
    Member:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        role: { type: string, enum: [owner, admin, member] }
...
    AllowlistEntry:
      type: object
      properties:
        email: { type: string, format: email }
        is_external: { type: boolean }
        claimed_at: { type: string, format: date-time, nullable: true }
        claimed_by: { type: string, format: uuid, nullable: true }
...
    Site:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        slug: { type: string }
        owner_id: { type: string, format: uuid }
        access_mode: { type: string, enum: [public, password, allowlist, org_only] }
        current_version_id: { type: string, format: uuid, nullable: true }
        live_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error: { type: string, description: Stable machine-readable code }
        message: { type: string }
    QuotaExceeded:
      type: object
      description: 402 body (see internal/quota.ExceededError); the dashboard opens the upgrade/sales modal.
      properties:
        limit: { type: string, enum: [sites_per_user, members_per_org] }
        current: { type: integer, format: int64 }
        max: { type: integer, format: int64 }
        plan_tier: { type: string }
        next_tier: { type: string }
        upgrade_url: { type: string, format: uri }
        sales_url: { type: string, format: uri }
  responses:
    Unauthorized:
      description: Missing/invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted (e.g. not admin/owner, expired link, not on allowlist, external sharing disabled)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Malformed request / reserved slug / verification failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource absent (or invisible under the caller's tenant)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Slug/host already in use by another org/site (global host registry)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    QuotaExceeded:
      description: A hard cap was crossed (cloud build only)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/QuotaExceeded" }